+++ app/controllers/steps/opening/done_controller.rb
@@ -1,0 +1,7 @@
+module Steps
+  module Opening
+    class DoneController < Steps::OpeningStepController
+      def show; end
+    end
+  end
+end
