+++ app/controllers/steps/opening/error_but_continue_controller.rb
@@ -1,0 +1,9 @@
+module Steps
+  module Opening
+    class ErrorButContinueController < Steps::OpeningStepController
+      def show
+        @courtfinder_ok = C100App::CourtfinderAPI.new.is_ok?
+      end
+    end
+  end
+end
