+++ app/controllers/steps/opening/no_court_found_controller.rb
@@ -1,0 +1,7 @@
+module Steps
+  module Opening
+    class NoCourtFoundController < Steps::OpeningStepController
+      def show; end
+    end
+  end
+end
