+++ app/controllers/steps/opening/warning_controller.rb
@@ -1,0 +1,9 @@
+module Steps
+  module Opening
+    class WarningController < Steps::OpeningStepController
+      skip_before_action :update_navigation_stack
+
+      def show; end
+    end
+  end
+end
