--- app/controllers/steps/screener/done_controller.rb
@@ -1,7 +1,0 @@
-module Steps
-  module Screener
-    class DoneController < Steps::ScreenerStepController
-      def show; end
-    end
-  end
-end
