--- app/controllers/steps/screener/error_but_continue_controller.rb
@@ -1,9 +1,0 @@
-module Steps
-  module Screener
-    class ErrorButContinueController < Steps::ScreenerStepController
-      def show
-        @courtfinder_ok = C100App::CourtfinderAPI.new.is_ok?
-      end
-    end
-  end
-end
