--- app/controllers/steps/screener/no_court_found_controller.rb
@@ -1,7 +1,0 @@
-module Steps
-  module Screener
-    class NoCourtFoundController < Steps::ScreenerStepController
-      def show; end
-    end
-  end
-end
