--- app/controllers/steps/screener/warning_controller.rb
@@ -1,9 +1,0 @@
-module Steps
-  module Screener
-    class WarningController < Steps::ScreenerStepController
-      skip_before_action :update_navigation_stack
-
-      def show; end
-    end
-  end
-end
