+++ app/forms/steps/opening/postcode_form.rb
@@ -1,0 +1,19 @@
+module Steps
+  module Opening
+    class PostcodeForm < BaseForm
+      attribute :children_postcode, StrippedString
+      validates :children_postcode, presence: true, full_uk_postcode: true
+
+      private
+
+      def persist!
+        raise C100ApplicationNotFound unless c100_application
+
+        c100_application.update(
+          children_postcode: children_postcode,
+          court: nil, # reset court attribute, so we start fresh
+        )
+      end
+    end
+  end
+end
