--- app/forms/steps/screener/postcode_form.rb
@@ -1,19 +1,0 @@
-module Steps
-  module Screener
-    class PostcodeForm < BaseForm
-      attribute :children_postcode, StrippedString
-      validates :children_postcode, presence: true, full_uk_postcode: true
-
-      private
-
-      def persist!
-        raise C100ApplicationNotFound unless c100_application
-
-        c100_application.update(
-          children_postcode: children_postcode,
-          court: nil, # reset court attribute, so we start fresh
-        )
-      end
-    end
-  end
-end
