+++ app/views/steps/miam_exemptions/exit_page/show.en.html.erb
@@ -30,7 +30,7 @@
       <li>Ask the mediator for a document confirming your attendance</li>
     </ol>
 
-    <p class="govuk-body"><%=t 'steps.shared.screener_exit_actions.cait_info_html' %></p>
+    <p class="govuk-body"><%=t '.cait_info_html' %></p>
 
     <%= render partial: 'steps/shared/kickout_survey' %>
   </div>
