+++ app/views/steps/opening/done/show.en.html.erb
@@ -1,0 +1,14 @@
+<% title 'Eligible to apply' %>
+<% step_header %>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+    <h1 class="govuk-heading-xl">You’re eligible to apply online</h1>
+
+    <div class="govuk-inset-text">
+      <p>If you’re a solicitor representing a client, complete this application on their behalf.</p>
+    </div>
+
+    <%= link_button 'Continue', entrypoint_v1_path %>
+  </div>
+</div>
