+++ app/views/steps/opening/start/show.en.html.erb
@@ -1,0 +1,29 @@
+<% title '' %>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+    <h1 class="govuk-heading-xl">
+      Apply to court about child arrangements
+    </h1>
+
+    <p class="govuk-body-l">Use this service to make an application for child arrangements in England or Wales.</p>
+
+    <p class="govuk-body">Your application will be sent to a court in the area where the child or children live.</p>
+
+    <div class="govuk-grid-row govuk-!-margin-top-6">
+      <div class="govuk-grid-column-one-half">
+        <%= link_to edit_steps_opening_postcode_path, class: 'govuk-button govuk-button--start', data: { module: 'govuk-button' }, role: 'button', draggable: false do %>
+          Check eligibility
+          <svg class="govuk-button__start-icon" xmlns="http://www.w3.org/2000/svg" width="17.5" height="19" viewBox="0 0 33 40" aria-hidden="true" focusable="false">
+            <path fill="currentColor" d="M0 0h13l20 20-20 20H0l20-20z" />
+          </svg>
+        <% end %>
+      </div>
+
+      <div class="govuk-grid-column-one-half govuk-!-margin-top-2">
+        <%= link_to 'Or return to a saved application', new_user_session_path,
+                    class: 'govuk-link govuk-link--no-visited-state govuk-!-font-size-19' %>
+      </div>
+    </div>
+  </div>
+</div>
