+++ app/views/steps/opening/warning/show.html.erb
