--- app/views/steps/screener/no_court_found/show.en.html.erb
@@ -1,22 +1,0 @@
-<% title 'Not eligible' %>
-<% step_header %>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-xl">Sorry, you cannot apply online</h1>
-
-    <p class="govuk-body-l">This service is only available in England and Wales.</p>
-
-    <p class="govuk-body">
-      If the child or children live in Scotland or Northern Ireland, you will not be able to use this service to submit
-      your application.
-    </p>
-
-    <p class="govuk-body">
-      If you think there’s been an error finding an appropriate court, you
-      can <%= link_to 'report the problem', about_contact_path, class: 'govuk-link ga-pageLink', data: { ga_category: 'no court found', ga_label: 'contact' } %>.
-    </p>
-
-    <%= render partial: '/steps/shared/screener_exit_actions' %>
-  </div>
-</div>
