--- app/views/steps/screener/postcode/edit.html.erb
@@ -1,29 +1,0 @@
-<% title t('.page_title') %>
-<% step_header %>
-
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <%= govuk_error_summary %>
-
-    <h1 class="govuk-heading-xl"><%=t '.heading' %></h1>
-
-    <p class="govuk-body-l"><%=t '.lead_text' %></p>
-
-    <%= step_form @form_object do |f| %>
-      <%= f.govuk_text_field :children_postcode, width: 'one-third', autocomplete: 'postal-code' %>
-
-      <%= f.continue_button %>
-    <% end %>
-
-    <details class="govuk-details" data-module="govuk-details">
-      <summary class="govuk-details__summary">
-        <span class="govuk-details__summary-text" data-ga-category="explanations" data-ga-label="screener postcode">
-          <%=t '.details.summary' %>
-        </span>
-      </summary>
-      <div class="govuk-details__text">
-        <%=t '.details.text_html' %>
-      </div>
-    </details>
-  </div>
-</div>
