--- app/views/steps/shared/_screener_exit_actions.html.erb
@@ -1,7 +1,0 @@
-<h2 class="govuk-heading-l"><%= t '.what_to_do_now.heading' %></h2>
-
-<p class="govuk-body"><%= t '.what_to_do_now.copy' %></p>
-
-<%= link_button t('.download_c100_link'), 'https://formfinder.hmctsformfinder.justice.gov.uk/c100-eng.pdf', class: 'ga-pageLink', data: { module: 'govuk-button', ga_category: 'screener kickout', ga_label: 'paper form' } %>
-
-<p class="govuk-body"><%=t '.cait_info_html' %></p>
