+++ spec/controllers/steps/miam/consent_order_controller_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe Steps::Miam::ConsentOrderController, type: :controller do
   it_behaves_like 'a controller that checks the application payment status', for_action: :edit
   it_behaves_like 'a savepoint step controller'
 
-  # Following examples will stub the screener sanity check for simplicity,
+  # Following examples will stub the court sanity check for simplicity,
   # as the above `savepoint` examples already test the involved code.
 
   it_behaves_like 'an intermediate step controller', Steps::Miam::ConsentOrderForm, C100App::MiamDecisionTree do
