+++ spec/controllers/steps/opening/done_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe Steps::Opening::DoneController, type: :controller do
+  it_behaves_like 'a show step controller'
+end
