--- spec/controllers/steps/screener/no_court_found_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe Steps::Screener::NoCourtFoundController, type: :controller do
-  it_behaves_like 'a show step controller'
-end
